lambda/ConvertPdfToTxt/requirements.txt
pdfminer.six==20191110
pycryptodome
six
requests
boto3
sortedcontainers
chardet
